requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
python-multipart==0.0.20
psycopg2-binary==2.9.10
sqlalchemy==2.0.34
celery==5.3.4
redis==5.0.1
minio==7.2.0
Pillow==11.3.0
python-json-logger==2.0.7
psutil==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
